Разделяет на модули, добавляет генератор уникальных идентификаторов, меняет функцию CreateAccomodations
